+++ src/emulator/controller.ts
@@ -8,19 +8,20 @@ import * as logger from "../logger";
 import * as track from "../track";
 import * as utils from "../utils";
 import { getCredentialPathAsync } from "../defaultCredentials";
-import { EmulatorRegistry } from "../emulator/registry";
+import { EmulatorRegistry } from "./registry";
 import {
   Address,
   ALL_SERVICE_EMULATORS,
   EmulatorInstance,
   Emulators,
   EMULATORS_SUPPORTED_BY_UI,
-} from "../emulator/types";
-import { Constants, FIND_AVAILBLE_PORT_BY_DEFAULT } from "../emulator/constants";
-import { FunctionsEmulator } from "../emulator/functionsEmulator";
-import { DatabaseEmulator, DatabaseEmulatorArgs } from "../emulator/databaseEmulator";
-import { FirestoreEmulator, FirestoreEmulatorArgs } from "../emulator/firestoreEmulator";
-import { HostingEmulator } from "../emulator/hostingEmulator";
+  isEmulator,
+} from "./types";
+import { Constants, FIND_AVAILBLE_PORT_BY_DEFAULT } from "./constants";
+import { FunctionsEmulator } from "./functionsEmulator";
+import { DatabaseEmulator, DatabaseEmulatorArgs } from "./databaseEmulator";
+import { FirestoreEmulator, FirestoreEmulatorArgs } from "./firestoreEmulator";
+import { HostingEmulator } from "./hostingEmulator";
 import { FirebaseError } from "../error";
 import * as getProjectId from "../getProjectId";
 import { PubsubEmulator } from "./pubsubEmulator";
@@ -32,6 +33,11 @@ import { LoggingEmulator } from "./loggingEmulator";
 import * as dbRulesConfig from "../database/rulesConfig";
 import { EmulatorLogger } from "./emulatorLogger";
 import * as portUtils from "./portUtils";
+import { EmulatorHubClient } from "./hubClient";
+import { promptOnce } from "../prompt";
+import * as rimraf from "rimraf";
+import { FLAG_EXPORT_ON_EXIT_NAME } from "./commandUtils";
+import { fileExistsSync } from "../fsutils";
 
 async function getAndCheckAddress(emulator: Emulators, options: any): Promise<Address> {
   const host = Constants.normalizeHost(
@@ -48,13 +54,13 @@ async function getAndCheckAddress(emulator: Emulators, options: any): Promise<Ad
     findAvailablePort = FIND_AVAILBLE_PORT_BY_DEFAULT[emulator];
   }
 
-  const logger = EmulatorLogger.forEmulator(emulator);
+  const loggerForEmulator = EmulatorLogger.forEmulator(emulator);
   const portOpen = await portUtils.checkPortOpen(port, host);
   if (!portOpen) {
     if (findAvailablePort) {
       const newPort = await portUtils.findAvailablePort(host, port);
       if (newPort != port) {
-        logger.logLabeled(
+        loggerForEmulator.logLabeled(
           "WARN",
           emulator,
           `${Constants.description(
@@ -66,12 +72,12 @@ async function getAndCheckAddress(emulator: Emulators, options: any): Promise<Ad
     } else {
       await cleanShutdown();
       const description = Constants.description(emulator);
-      logger.logLabeled(
+      loggerForEmulator.logLabeled(
         "WARN",
         emulator,
         `Port ${port} is not open on ${host}, could not start ${description}.`
       );
-      logger.logLabeled(
+      loggerForEmulator.logLabeled(
         "WARN",
         emulator,
         `To select a different host/port, specify that host/port in a firebase.json config file:
@@ -91,7 +97,7 @@ async function getAndCheckAddress(emulator: Emulators, options: any): Promise<Ad
 
   if (portUtils.isRestricted(port)) {
     const suggested = portUtils.suggestUnrestricted(port);
-    logger.logLabeled(
+    loggerForEmulator.logLabeled(
       "WARN",
       emulator,
       `Port ${port} is restricted by some web browsers, including Chrome. You may want to choose a different port such as ${suggested}.`
@@ -110,20 +116,45 @@ export async function startEmulator(instance: EmulatorInstance): Promise<void> {
   await EmulatorRegistry.start(instance);
 }
 
+/**
+ * Exports emulator data on clean exit (SIGINT or process end)
+ * @param options
+ */
+export async function exportOnExit(options: any) {
+  const exportOnExitDir = options.exportOnExit;
+  if (exportOnExitDir) {
+    try {
+      utils.logBullet(
+        `Automatically exporting data using ${FLAG_EXPORT_ON_EXIT_NAME} "${exportOnExitDir}" ` +
+          "please wait for the export to finish..."
+      );
+      await exportEmulatorData(exportOnExitDir, options);
+    } catch (e) {
+      utils.logWarning(e);
+      utils.logWarning(`Automatic export to "${exportOnExitDir}" failed, going to exit now...`);
+    }
+  }
+}
+
+/**
+ * Hook to do things when we're exiting cleanly (this does not include errors). Will be skipped on a second SIGINT
+ * @param options
+ */
+export async function onExit(options: any) {
+  await exportOnExit(options);
+}
+
+/**
+ * Hook to clean up on shutdown (includes errors). Will be skipped on a third SIGINT
+ * Stops all running emulators in parallel.
+ */
 export async function cleanShutdown(): Promise<void> {
   EmulatorLogger.forEmulator(Emulators.HUB).logLabeled(
     "BULLET",
     "emulators",
     "Shutting down emulators."
   );
-  for (const name of EmulatorRegistry.listRunning()) {
-    EmulatorLogger.forEmulator(name).logLabeled(
-      "BULLET",
-      name,
-      `Stopping ${Constants.description(name)}`
-    );
-    await EmulatorRegistry.stop(name);
-  }
+  await EmulatorRegistry.stopAll();
 }
 
 export function filterEmulatorTargets(options: any): Emulators[] {
@@ -217,13 +248,24 @@ export async function startAll(options: any, noUi: boolean = false): Promise<voi
     const ignored = _.difference(requested, targets);
 
     for (const name of ignored) {
-      EmulatorLogger.forEmulator(name as Emulators).logLabeled(
-        "WARN",
-        name,
-        `Not starting the ${clc.bold(name)} emulator, make sure you have run ${clc.bold(
-          "firebase init"
-        )}.`
-      );
+      if (isEmulator(name)) {
+        EmulatorLogger.forEmulator(name as Emulators).logLabeled(
+          "WARN",
+          name,
+          `Not starting the ${clc.bold(name)} emulator, make sure you have run ${clc.bold(
+            "firebase init"
+          )}.`
+        );
+      } else {
+        // this should not work:
+        // firebase emulators:start --only doesnotexit
+        throw new FirebaseError(
+          `${name} is not a valid emulator name, valid options are: ${JSON.stringify(
+            ALL_SERVICE_EMULATORS
+          )}`,
+          { exit: 1 }
+        );
+      }
     }
   }
 
@@ -238,10 +280,26 @@ export async function startAll(options: any, noUi: boolean = false): Promise<voi
     version: "unknown",
   };
   if (options.import) {
-    const importDir = path.resolve(options.import);
-    exportMetadata = JSON.parse(
-      fs.readFileSync(path.join(importDir, HubExport.METADATA_FILE_NAME), "utf8").toString()
-    ) as ExportMetadata;
+    const importFilePath = path.join(path.resolve(options.import), HubExport.METADATA_FILE_NAME);
+    if (fileExistsSync(importFilePath)) {
+      exportMetadata = JSON.parse(
+        fs
+          .readFileSync(
+            path.join(path.resolve(options.import), HubExport.METADATA_FILE_NAME),
+            "utf8"
+          )
+          .toString()
+      ) as ExportMetadata;
+    } else {
+      // could happen when an export was interrupted mid-export...
+      EmulatorLogger.forEmulator(Emulators.HUB).logLabeled(
+        "WARN",
+        "emulators",
+        `Import/Export metadata file does not exist. ${clc.bold(
+          "Skipping data import!"
+        )} Metadata file location: ${importFilePath}`
+      );
+    }
   }
 
   if (shouldStart(options, Emulators.FUNCTIONS)) {
@@ -520,4 +578,90 @@ export async function startAll(options: any, noUi: boolean = false): Promise<voi
       await instance.connect();
     }
   }
+}
+
+/**
+ * Exports data from emulators that support data export. Used with `emulators:export` and with the --export-on-exit flag.
+ * @param exportPath
+ * @param options
+ */
+export async function exportEmulatorData(exportPath: string, options: any) {
+  const projectId = options.project;
+  if (!projectId) {
+    throw new FirebaseError(
+      "Could not determine project ID, make sure you're running in a Firebase project directory or add the --project flag.",
+      { exit: 1 }
+    );
+  }
+
+  const hubClient = new EmulatorHubClient(projectId);
+  if (!hubClient.foundHub()) {
+    throw new FirebaseError(
+      `Did not find any running emulators for project ${clc.bold(projectId)}.`,
+      { exit: 1 }
+    );
+  }
+
+  try {
+    await hubClient.getStatus();
+  } catch (e) {
+    const filePath = EmulatorHub.getLocatorFilePath(projectId);
+    throw new FirebaseError(
+      `The emulator hub for ${projectId} did not respond to a status check. If this error continues try shutting down all running emulators and deleting the file ${filePath}`,
+      { exit: 1 }
+    );
+  }
+
+  utils.logBullet(
+    `Found running emulator hub for project ${clc.bold(projectId)} at ${hubClient.origin}`
+  );
+
+  // If the export target directory does not exist, we should attempt to create it
+  const exportAbsPath = path.resolve(exportPath);
+  if (!fs.existsSync(exportAbsPath)) {
+    utils.logBullet(`Creating export directory ${exportAbsPath}`);
+    fs.mkdirSync(exportAbsPath);
+  }
+
+  // Check if there is already an export there and prompt the user about deleting it
+  const existingMetadata = HubExport.readMetadata(exportAbsPath);
+  if (existingMetadata && !(options.force || options.exportOnExit)) {
+    if (options.noninteractive) {
+      throw new FirebaseError(
+        "Export already exists in the target directory, re-run with --force to overwrite.",
+        { exit: 1 }
+      );
+    }
+
+    const prompt = await promptOnce({
+      type: "confirm",
+      message: `The directory ${exportAbsPath} already contains export data. Exporting again to the same directory will overwrite all data. Do you want to continue?`,
+      default: false,
+    });
+
+    if (!prompt) {
+      throw new FirebaseError("Command aborted", { exit: 1 });
+    }
+  }
+
+  // Remove all existing data (metadata.json will be overwritten automatically)
+  if (existingMetadata) {
+    if (existingMetadata.firestore) {
+      const firestorePath = path.join(exportAbsPath, existingMetadata.firestore.path);
+      utils.logBullet(`Deleting directory ${firestorePath}`);
+      rimraf.sync(firestorePath);
+    }
+  }
+
+  utils.logBullet(`Exporting data to: ${exportAbsPath}`);
+  try {
+    await hubClient.postExport(exportAbsPath);
+  } catch (e) {
+    throw new FirebaseError("Export request failed, see emulator logs for more information.", {
+      exit: 1,
+      original: e,
+    });
+  }
+
+  utils.logSuccess("Export complete");
 }