+++ CHANGELOG.md
@@ -1,1 +1,2 @@
 - Fixes an issue where the suggested redeploy command for Firebase Functions was incorrect for names with dashes.
+- Adds a the `--export-on-exit` flag to `emulators:start` and `emulators:exec` to automatically export emulator data on command exit (#2224)